+++ Tests/SwiftLintFrameworkTests/AutomaticRuleTests.generated.swift
@@ -516,6 +516,12 @@ class PatternMatchingKeywordsRuleTests: XCTestCase {
     }
 }
 
+class PhohibitedNaNComparisonRuleTests: XCTestCase {
+    func testWithDefaultConfiguration() {
+        verifyRule(PhohibitedNaNComparisonRule.description)
+    }
+}
+
 class PreferSelfTypeOverTypeOfSelfRuleTests: XCTestCase {
     func testWithDefaultConfiguration() {
         verifyRule(PreferSelfTypeOverTypeOfSelfRule.description)